requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.5.3
asgiref==3.8.1
attrs==25.1.0
azure-core==1.38.2
azure-storage-blob==12.28.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
dj-database-url==3.0.1
Django==5.1.6
django-cuser==2017.3.16
django-elasticsearch-dsl==8.0
django-environ==0.12.0
django-modeltranslation==0.19.17
django-parler==2.3
django-storages==1.14.6
django-tinymce==4.1.0
dnspython==2.7.0
docker==7.1.0
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.19.1
elastic-transport==8.17.0
elasticsearch==8.17.1
elasticsearch-dsl==8.17.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
orjson==3.10.16
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.1
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.1
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.1
pyxdg==0.28
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.22.3
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.45
sqlparse==0.5.3
starlette==0.46.1
text-unidecode==1.3
timeago==1.0.16
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
ujson==5.10.0
urllib3==2.3.0
userpath==1.9.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
whitenoise==6.9.0
